feat: add Notification Settings, Support, GPA Hub, QR Scan, and Student Communication pages
- Implemented NotificationSettingsPage for managing user notification preferences.
- Created SupportPage with Help Center and My Feedback tabs for user assistance.
- Added GPAHubPage with GPA Predictor and Calculator functionalities.
- Developed QRScanPage for scanning and verifying QR codes.
- Introduced StudentCommunicationPage for accessing notifications, announcements, and notices.

+++ backend/internal/api/predictions.go
@@ -12,7 +12,7 @@ import (
 type GPAPredictionItem struct {
 	CourseCode  string  `json:"course_code"`
 	Credits     int32   `json:"credits"`
-	TotalScore  int32   `json:"total_score"`
+	TotalScore  float64 `json:"total_score"`
 	GradePoints float64 `json:"grade_points"`
 	GradeLetter string  `json:"grade_letter"`
 }
